feat: move lc/institution enforcement functionality and tests to specialized service class
