+++ src/main/java/org/broadinstitute/consent/http/ConsentModule.java
@@ -62,6 +62,7 @@ import org.broadinstitute.consent.http.service.OidcService;
 import org.broadinstitute.consent.http.service.OntologyService;
 import org.broadinstitute.consent.http.service.SupportRequestService;
 import org.broadinstitute.consent.http.service.UseRestrictionConverter;
+import org.broadinstitute.consent.http.service.UserEnforcementService;
 import org.broadinstitute.consent.http.service.UserService;
 import org.broadinstitute.consent.http.service.VoteService;
 import org.broadinstitute.consent.http.service.dao.DaaServiceDAO;
@@ -607,7 +608,15 @@ public class ConsentModule extends AbstractModule {
         providesDraftService(),
         providesInstitutionService(),
         providesDACAutomationRuleDAO(),
-        providesDatasetAuthorizationReaderDAO());
+        providesDatasetAuthorizationReaderDAO(),
+        providesUserEnforcementService());
+  }
+
+  @Provides
+  UserEnforcementService providesUserEnforcementService() {
+    return new UserEnforcementService(
+        providesJdbi(),
+        providesUserServiceDAO());
   }
 
   @Provides
