+++ src/test/java/org/broadinstitute/consent/http/service/UserEnforcementServiceTest.java
@@ -1,0 +1,302 @@
+package org.broadinstitute.consent.http.service;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import java.util.stream.Stream;
+import org.broadinstitute.consent.http.AbstractTestHelper;
+import org.broadinstitute.consent.http.db.InstitutionDAO;
+import org.broadinstitute.consent.http.db.LibraryCardDAO;
+import org.broadinstitute.consent.http.db.UserDAO;
+import org.broadinstitute.consent.http.models.Institution;
+import org.broadinstitute.consent.http.models.LibraryCard;
+import org.broadinstitute.consent.http.models.User;
+import org.broadinstitute.consent.http.service.dao.UserServiceDAO;
+import org.jdbi.v3.core.Jdbi;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class UserEnforcementServiceTest extends AbstractTestHelper {
+
+  @Mock
+  private InstitutionDAO institutionDAO;
+  @Mock
+  private LibraryCardDAO libraryCardDAO;
+  @Mock
+  private UserDAO userDAO;
+  @Mock
+  private UserServiceDAO userServiceDAO;
+  @Mock
+  private Jdbi jdbi;
+
+  private UserEnforcementService service;
+
+  @BeforeEach
+  void setUp() {
+    when(jdbi.onDemand(InstitutionDAO.class)).thenReturn(institutionDAO);
+    when(jdbi.onDemand(LibraryCardDAO.class)).thenReturn(libraryCardDAO);
+    when(jdbi.onDemand(UserDAO.class)).thenReturn(userDAO);
+    service = new UserEnforcementService(jdbi, userServiceDAO);
+  }
+
+  @Test
+  void hasLibraryCard() {
+    User testUser = generateUser();
+    testUser.setLibraryCard(new LibraryCard());
+    assertTrue(service.hasLibraryCard(testUser));
+  }
+
+  @Test
+  void hasLibraryCard_NoLibraryCard() {
+    User testUser = generateUser();
+    assertFalse(service.hasLibraryCard(testUser));
+  }
+
+  @Test
+  void hasMatchingInstitutionInDatabase() {
+    Institution institution = new Institution();
+    institution.setId(1);
+    assertTrue(service.hasMatchingInstitutionInDatabase(institution, institution));
+  }
+
+  @Test
+  void hasMatchingInstitutionInDatabase_NoInstitutionInDB() {
+    Institution institution = new Institution();
+    institution.setId(1);
+    assertFalse(service.hasMatchingInstitutionInDatabase(null, institution));
+  }
+
+  @Test
+  void hasMatchingInstitutionInDatabase_NoInstitutionEmailMap() {
+    Institution institution = new Institution();
+    institution.setId(1);
+    assertFalse(service.hasMatchingInstitutionInDatabase(institution, null));
+  }
+
+  @Test
+  void handleUserWithInstitutionInMap() {
+    User testUser = generateUser();
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitutionId(1);
+    assertFalse(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+  }
+
+  @Test
+  void handleUserWithInstitutionInMap_DifferentInDatabase() {
+    User testUser = generateUser();
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitutionId(2);
+    assertTrue(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+    verify(userDAO).updateInstitutionId(testUser.getUserId(), institutionFromEmail.getId());
+  }
+
+  @Test
+  void handleUserWithInstitutionInMap_DifferentInDatabaseWithLibraryCard() {
+    User testUser = generateUser();
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitution(institutionFromEmail);
+    Institution institutionFromDatabase = new Institution();
+    institutionFromDatabase.setId(2);
+    testUser.setInstitutionId(institutionFromDatabase.getId());
+    String soDomain = service.trimmedEmailDomain(signingOfficial.getEmail());
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(signingOfficial);
+    when(institutionDAO.findInstitutionByDomain(soDomain)).thenReturn(institutionFromDatabase);
+
+    assertTrue(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+    verify(userServiceDAO).updateInstitutionAndClearLibraryCardForUser(testUser.getUserId(), institutionFromEmail.getId());
+  }
+
+  @Test
+  void handleUserWithInstitutionInMap_DifferentInDatabaseWithLibraryCard_SO_NFE() {
+    User testUser = generateUser();
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitution(institutionFromEmail);
+    Institution institutionFromDatabase = new Institution();
+    institutionFromDatabase.setId(2);
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(null);
+
+    assertTrue(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+    verify(userServiceDAO).updateInstitutionAndClearLibraryCardForUser(testUser.getUserId(), institutionFromEmail.getId());
+  }
+
+  @Test
+  void handleUserWithInstitutionInMap_SameInDatabaseWithLC() {
+    User testUser = generateUser();
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitution(institutionFromEmail);
+    testUser.setInstitutionId(1);
+    String soDomain = service.trimmedEmailDomain(signingOfficial.getEmail());
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(signingOfficial);
+    when(institutionDAO.findInstitutionByDomain(soDomain)).thenReturn(institutionFromEmail);
+
+    assertFalse(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+  }
+
+
+  @Test
+  void handleUserWithInstitutionInMap_SameInDatabaseWithLCFromDifferentOrg() {
+    User testUser = generateUser();
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitution(institutionFromEmail);
+    testUser.setInstitutionId(1);
+    String soDomain = service.trimmedEmailDomain(signingOfficial.getEmail());
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(signingOfficial);
+    when(institutionDAO.findInstitutionByDomain(soDomain)).thenReturn(null);
+
+    assertTrue(service.handleUserWithInstitutionInMap(testUser, institutionFromEmail.getId()));
+    verify(userDAO, times(0)).updateInstitutionId(any(), any());
+    verify(libraryCardDAO).deleteAllLibraryCardsByUser(testUser.getUserId());
+  }
+
+  @Test
+  void needsLibraryCardRemovedForUser() {
+    User testUser = generateUser();
+    Institution institution = new Institution();
+    assertFalse(service.needsLibraryCardRemovedForUser(testUser, institution.getId()));
+  }
+
+  @Test
+  void needsLibraryCardRemovedForUser_SO_NFE() {
+    User testUser = generateUser();
+    Institution institution = new Institution();
+    institution.setId(testUser.getInstitutionId());
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(null);
+    assertTrue(service.needsLibraryCardRemovedForUser(testUser, institution.getId()));
+  }
+
+  @Test
+  void needsLibraryCardRemovedForUser_SO_DifferentInstitution() {
+    User testUser = generateUser();
+    User signingOfficial = generateUser();
+    LibraryCard lc = new LibraryCard();
+    lc.setCreateUserId(signingOfficial.getUserId());
+    testUser.setLibraryCard(lc);
+    Institution institutionFromEmail = new Institution();
+    institutionFromEmail.setId(1);
+    testUser.setInstitution(institutionFromEmail);
+    String soDomain = service.trimmedEmailDomain(signingOfficial.getEmail());
+    Institution soInstitution = new Institution();
+    soInstitution.setId(2);
+
+    when(userDAO.findUserById(signingOfficial.getUserId())).thenReturn(signingOfficial);
+    when(institutionDAO.findInstitutionByDomain(soDomain)).thenReturn(soInstitution);
+    assertTrue(service.needsLibraryCardRemovedForUser(testUser, institutionFromEmail.getId()));
+  }
+
+  public static Stream<Arguments> testEnforceInstitutionAndLibraryCardVariations() {
+    User testUser = generateUser();
+    Institution institution1 = new Institution();
+    institution1.setId(1);
+    Institution institution2 = new Institution();
+    institution2.setId(2);
+    LibraryCard libraryCards1 = new LibraryCard();
+    LibraryCard libraryCard2 = new LibraryCard();
+    libraryCard2.setCreateUserId(1);
+    return Stream.of(
+        Arguments.of(institution1, testUser, libraryCards1, true),
+        Arguments.of(institution2, testUser, libraryCards1, true),
+        Arguments.of(institution1, testUser, libraryCard2, false),
+        Arguments.of(institution1, testUser, null, false),
+        Arguments.of(null, testUser, libraryCards1, true),
+        Arguments.of(null, testUser, null, false)
+    );
+  }
+
+  @ParameterizedTest
+  @MethodSource
+  void testEnforceInstitutionAndLibraryCardVariations(
+      Institution institutionFromMap, User testUser, LibraryCard card, boolean expectsUserMod) {
+    testUser.setLibraryCard(card);
+    User alteredUser = new User();
+    alteredUser.setEmail(testUser.getEmail());
+    String domain = service.trimmedEmailDomain(testUser.getEmail());
+    if (institutionFromMap != null) {
+      when(institutionDAO.findInstitutionIdByDomain(domain))
+          .thenReturn(institutionFromMap.getId());
+    } else {
+      when(institutionDAO.findInstitutionIdByDomain(domain))
+          .thenReturn(null);
+    }
+    if (expectsUserMod) {
+      when(userDAO.findUserByEmail(testUser.getEmail())).thenReturn(testUser, alteredUser);
+      validateAlteredUserIsReturned(
+          testUser, service.enforceInstitutionAndLibraryCardRules(testUser.getEmail()));
+    } else {
+      when(userDAO.findUserByEmail(testUser.getEmail())).thenReturn(testUser);
+      validateUserIsUnmodified(
+          testUser, service.enforceInstitutionAndLibraryCardRules(testUser.getEmail()));
+    }
+  }
+
+  private void validateUserIsUnmodified(User left, User right) {
+    assertEquals(left.getEmail(), right.getEmail());
+    assertEquals(left.getInstitutionId(), right.getInstitutionId());
+    assertEquals(left.getLibraryCard(), right.getLibraryCard());
+    assertEquals(left.getInstitutionId(), right.getInstitutionId());
+  }
+
+  private void validateAlteredUserIsReturned(User left, User right) {
+    assertEquals(left.getEmail(), right.getEmail());
+    assertNotEquals(left.getInstitutionId(), right.getInstitutionId());
+  }
+
+  private static User generateUser() {
+    User u = new User();
+    int i1 = randomInt(10, 50);
+    int i2 = randomInt(10, 50);
+    int i3 = randomInt(5, 25);
+    String email = randomAlphabetic(i1) + "@" + randomAlphabetic(i2) + "." + randomAlphabetic(i3);
+    String displayName = randomAlphabetic(i1) + " " + randomAlphabetic(i2);
+    u.setEmail(email);
+    u.setEraCommonsId(email);
+    u.setDisplayName(displayName);
+    u.setUserId(randomInt(1, 100));
+    u.setInstitutionId(randomInt(1, 100));
+    return u;
+  }
+
+}
